CreateFinalStudentScore then Finshed SchoolYear And FinhedSemester and start new semsterin schoolyear

+++ QalamAndNoor/Models/Course.cs
@@ -9,6 +9,6 @@
         public bool IsEnriching { get; set; }
         public int ClassId { get; set; }
         public bool RequiredToPass { get; set; }
-        public int RequiredGradeToPass { get; set; }
+        public double RequiredGradeToPass { get; set; }
     }
 }
